database/requirements.txt
# Database requirements for Physics Assistant
# Core database drivers
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.14.1

# Environment and configuration
python-dotenv==1.0.0

# Database utilities and migration tools
alembic==1.13.1
sqlalchemy==2.0.23

# Connection pooling and utilities
aiopg==1.4.0
asyncpg-utils==0.2.1

# API and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Monitoring and metrics
prometheus_client==0.19.0

# Document Processing Pipeline Dependencies (Phase 3.2)
# Mathematical processing
sympy>=1.12
numpy>=1.24.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Document processing
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11

# Natural language processing (optional)
# spacy>=3.6.0
# Note: For spaCy, install model with: python -m spacy download en_core_web_sm

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Backup and compression utilities (system dependencies required)
# Note: pg_dump, pg_restore, psql must be installed separately